pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiautomouse"
version = "0.1.0"
description = "Windows desktop automation MVP with deterministic YAML/JSON macros."
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
  "easyocr>=1.7.2",
  "mss>=9.0.1",
  "numpy>=1.26.4",
  "opencv-python>=4.10.0.84",
  "Pillow>=10.4.0",
  "playwright>=1.52.0",
  "PyInstaller>=6.13.0",
  "pydantic>=2.8.2",
  "pywinauto>=0.6.8",
  "PyYAML>=6.0.2",
  "PySide6>=6.8.1",
  "pytesseract>=0.3.13",
  "python-json-logger>=2.0.7",
  "winsdk>=1.0.0b10",
]

[project.optional-dependencies]
dev = [
  "PyInstaller>=6.13.0",
  "pytest>=8.3.3",
]

[project.scripts]
aiautomouse = "aiautomouse.cli:main"
aiautomouse-gui = "aiautomouse.gui.app:launch_gui"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
  "integration: Windows integration tests that may open desktop applications",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
